+++ app/controllers/api/web/episodes/magazine_controller.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Api
+  module Web
+    module Episodes
+      class MagazineController < ApplicationController
+        def show
+          magazine = Magazine.find(params[:id])
+          render json: Resource::Magazine.new(magazine)
+        end
+      end
+    end
+  end
+end
