--- app/controllers/episodes/magazine_controller.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Episodes
-  class MagazineController < ApplicationController
-    def show
-      magazine = Magazine.find(params[:id])
-      render json: Resource::Magazine.new(magazine)
-    end
-  end
-end
