--- app/controllers/series_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class SeriesController < ApplicationController
-  def index
-    render json: Series.all.includes(magazines: [:pages]).map(&Resource::Series.method(:new))
-  end
-end
