+++ spec/controllers/api/web/magazines_controller_create_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe MagazinesController, type: :controller do
+RSpec.describe Api::Web::MagazinesController, type: :controller do
   describe '#create' do
     subject do
       post :create, params: {
