+++ spec/controllers/api/web/magazines_controller_show_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe MagazinesController, type: :controller do
+RSpec.describe Api::Web::MagazinesController, type: :controller do
   subject do
     get :show, params: { id: magazine.id }
     response.body
