+++ spec/controllers/api/web/series_controller_index_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe SeriesController, type: :controller do
+RSpec.describe Api::Web::SeriesController, type: :controller do
   let!(:series) { create_list(:series, 3) }
   let!(:magazines) { create_list(:magazine, 10, series: series.first) }
